some more moving around
